# Squeeze more bytes out of slottedarray