--- a/src/Nethermind
+++ b/src/Nethermind
@@ -1,1 +1,0 @@
-/Users/aryan/Developer/nethermind-internship/nethermind