… src/Paprika.Tests/.vscode/settings.json
… src/Paprika.Tests/Data/SlottedArrayTests.Bla_Bla_Set_Get_Delete_Mid_Len_Key.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Defragment_when_no_more_space.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_Mid_lenKeys_odd=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_Mid_lenKeys_odd=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=0_lengthCutOff=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=0_lengthCutOff=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Hash_collisions.received.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Get_AnotherSet.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Mid_Len_Key.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Mid_Len_Key1_odd=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Mid_Len_Key1_odd=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Mid_Len_Keys_odd=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Mid_Len_Keys_odd=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Empty.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Temp_TBD_Hash_collisions.received.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Update_in_resize.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Update_in_situ.verified.bin
… src/Paprika.Tests/Paprika.Tests.sln
… src/Paprika.Tests/Values.cs
… src/Paprika/Data/Key.cs
… src/Paprika/Data/NibblePath.cs
… src/Paprika/Data/SlottedArray.cs
… src/Paprika/Merkle/Node.cs