@AryanGodara 2024-05-31
Changes Done: 
1. Add cases for length 5,6
2. Update PrepareKey() and UnprepareKey() to support these new types
3. Update ReadFrom() functions, to have another function for these cases which takes length and oddity as input (Can just pass in the preamble instead of having an int and a bool parameter)

Next steps: 
1. Update the WriteImpl and read methods, which set the first byte of trimmed for length and oddity metadata
2. Make sure all tests pass, and run benchmarks.
3. Work on the smaller keys: Possibly reduce hash size to 3 nibbles, and use preamble for length <= 6, and marker for length >=7