@AryanGodara 2024-06-06
Latest Benchmarks :-

Main Branch (Original) :-

| Method                              | sliceFrom | Mean         | Error      | StdDev     | Median       |
|------------------------------------ |---------- |-------------:|-----------:|-----------:|-------------:|
| Write_whole_page_of_data            | ?         | 14,300.79 ns | 172.493 ns | 378.627 ns | 14,151.02 ns |
| Read_existing_keys_prefix_different | ?         | 15,949.02 ns | 229.801 ns | 255.424 ns | 15,849.62 ns |
| Read_existing_keys_suffix_different | ?         | 16,024.92 ns | 132.107 ns | 103.141 ns | 16,008.06 ns |
| Read_nonexistent_keys               | ?         | 24,123.01 ns | 440.956 ns | 344.269 ns | 23,951.74 ns |
| Hash_collisions                     | ?         | 33,914.59 ns | 202.941 ns | 169.465 ns | 33,897.14 ns |
| EnumerateAll                        | ?         |  4,801.45 ns |  36.150 ns |  30.187 ns |  4,798.74 ns |
| Move_to_keys                        | ?         |    115.50 ns |   2.321 ns |   4.125 ns |    115.13 ns |
| UnPrepareKey                        | 0         |     49.24 ns |   0.288 ns |   0.225 ns |     49.17 ns |
| UnPrepareKey                        | 1         |     52.50 ns |   0.188 ns |   0.147 ns |     52.51 ns |
| UnPrepareKey                        | 62        |     18.62 ns |   0.097 ns |   0.086 ns |     18.59 ns |
| UnPrepareKey                        | 63        |     18.96 ns |   0.021 ns |   0.018 ns |     18.96 ns |
| UnPrepareKey                        | 64        |     17.88 ns |   0.059 ns |   0.049 ns |     17.87 ns |


This PR (After changes) :-

| Method                              | sliceFrom | Mean         | Error      | StdDev     |
|------------------------------------ |---------- |-------------:|-----------:|-----------:|
| Write_whole_page_of_data            | ?         | 14,737.75 ns | 291.854 ns | 273.000 ns |
| Read_existing_keys_prefix_different | ?         | 16,259.73 ns |  72.654 ns |  60.669 ns |
| Read_existing_keys_suffix_different | ?         | 16,541.15 ns | 311.198 ns | 345.896 ns |
| Read_nonexistent_keys               | ?         | 25,829.30 ns | 483.894 ns | 452.635 ns |
| Hash_collisions                     | ?         | 44,030.26 ns | 152.549 ns | 119.100 ns |
| EnumerateAll                        | ?         |  5,307.77 ns |  32.088 ns |  25.052 ns |
| Move_to_keys                        | ?         |    114.29 ns |   2.295 ns |   3.436 ns |
| UnPrepareKey                        | 0         |     50.57 ns |   1.018 ns |   0.795 ns |
| UnPrepareKey                        | 1         |     53.86 ns |   0.475 ns |   0.396 ns |
| UnPrepareKey                        | 62        |     18.47 ns |   0.027 ns |   0.023 ns |
| UnPrepareKey                        | 63        |     19.26 ns |   0.380 ns |   0.390 ns |
| UnPrepareKey                        | 64        |     17.92 ns |   0.072 ns |   0.064 ns |