@Scooletz 2024-09-04
Substituted by `KeyEncoding` at #375